requirements.txt
django==1.11.20
django-comments-xtd==2.0.9
django-contrib-comments==1.8.0
django-cors-headers==2.1.0
django-haystack==2.6.1
django-machina==0.6.0
django-mptt==0.8.7
django-reversion==2.0.10
django-reversion-compare==0.8.1
django-role-permissions==2.1.0
django-taggit==0.22.1
django-widget-tweaks==1.4.1
djangorestframework==3.6.4
gunicorn==19.7.1
mysqlclient==1.3.12
oauthlib==2.1.0
olefile==0.45.1
Pillow==4.3.0
pysolr==3.7.0
python-decouple==3.1
requests>=2.20.0
social-auth-app-django==2.1.0
django-notifications-hq==1.4.0
django-activity-stream==0.6.5
django-jsonfield==1.0.1
django-jsonfield-compat==0.4.4
elasticsearch==2.3.0
social-auth-core==1.7.0
-e git+https://github.com/devjones/PyEtherpadLite.git#egg=PyEtherpadLite
wiki==0.3.1
celery==4.2.1
future==0.17.1
django-categories==1.6
django-crispy-forms
